add about folder and fix font family bug
